lab2/requirements.txt
selenium==4.28.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
html2text==2024.2.26
lxml==5.3.0
lxml_html_clean
pdfplumber
pandas
matplotlib
seaborn
kagglehub